Uninstall @alextheman/utility from @neurosongs/utility and move parseQueryParameters to back-end
The listing of @alextheman/utility as a peer dependency of @neurosongs/utility was causing a version conflict because peer dependencies don't tend to get updated. @neurosongs/utility didn't really need the global utility as peer dependency either - it only needed it for APIError, which is also an indicator that it's a back-end only concern and should be placed there instead.
