+++ apps/back-end/src/utility/parseQueryParameter.ts
