+++ packages/utility/src/index.ts
@@ -1,6 +1,5 @@
 export { default as disableTscInDirectory } from "src/disableTscInDirectory";
 export { default as paginate } from "src/paginate";
 export { default as parseDate } from "src/parseDate";
-export { default as parseQueryParameter } from "src/parseQueryParameter";
 
 export type { PaginationResult } from "src/paginate";
